fix: Avoid 404 when loading an AppMap
This doesn't prevent the issue from occurring again in the future, but
it prevents the specific case where a 404 would cause future AppMaps not
to load. In older versions of the AppMap viewer, a missing web font was
referenced. The web view would attempt to load it, get a 404, and the
state would be stuck like this. This change only removes the dependency
on the missing web font.